# Gaud-370 - Detect ALL byte changes